requirements.txt
rwkv
--index-url https://download.pytorch.org/whl/cu117
torch
ninja

mido

# sudo apt-get install libasound-dev
# sudo apt-get install libjack-dev
python-rtmidi==1.4.*

# sudo apt-get install fluidsynth
pyFluidSynth

